refactor: reorganize notebook and quick insight components into dedicated subdirectories
